rename objects from checks to metrics
