+++ src/almanack/__init__.py
@@ -12,9 +12,9 @@ Foundation (NSF) via SHI under Grant No. 2327079.
 """
 
 from .book import read
-from .checks.data import get_table as table
-from .checks.data import process_repo_for_analysis
-from .checks.entropy.calculate_entropy import (
+from .metrics.data import get_table as table
+from .metrics.data import process_repo_for_analysis
+from .metrics.entropy.calculate_entropy import (
     calculate_aggregate_entropy,
     calculate_normalized_entropy,
 )
