+++ tests/metrics/entropy/test_calculate_entropy.py
@@ -4,7 +4,7 @@ Testing entropy functionality
 
 import pathlib
 
-from almanack.checks.entropy.calculate_entropy import (
+from almanack.metrics.entropy.calculate_entropy import (
     calculate_aggregate_entropy,
     calculate_normalized_entropy,
 )
